requirements.txt
customtkinter
pillow
opencv-python
matplotlib
ttkbootstrap
PyQt5